+++ docs/docs.go
@@ -972,6 +972,82 @@ const docTemplate = `{
                 }
             }
         },
+        "/api/admin/user/event/{event-id}/groups": {
+            "get": {
+                "security": [
+                    {
+                        "AppUserAuth": []
+                    }
+                ],
+                "description": "Updates the group mappings for an event with id",
+                "tags": [
+                    "Client"
+                ],
+                "operationId": "UpdateGroupMappingsEventID",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "APP",
+                        "name": "APP",
+                        "in": "header",
+                        "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Event ID",
+                        "name": "event-id",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/getPutAdminGroupIDsForEventIDRequestAndResponse"
+                        }
+                    }
+                }
+            }
+        },
+        "/api/admin/user/event/{event_id}/groups": {
+            "get": {
+                "security": [
+                    {
+                        "AppUserAuth": []
+                    }
+                ],
+                "description": "Get all group IDs where the current user is an admin",
+                "tags": [
+                    "Admin"
+                ],
+                "operationId": "AdminGetAdminGroupIDsForEventID",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "APP",
+                        "name": "APP",
+                        "in": "header",
+                        "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Event ID",
+                        "name": "event-id",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/getPutAdminGroupIDsForEventIDRequestAndResponse"
+                        }
+                    }
+                }
+            }
+        },
         "/api/admin/user/groups": {
             "get": {
                 "security": [
@@ -4630,6 +4706,17 @@ const docTemplate = `{
                 },
                 "web_url": {
                     "type": "string"
+                }
+            }
+        },
+        "getPutAdminGroupIDsForEventIDRequestAndResponse": {
+            "type": "object",
+            "properties": {
+                "group_ids": {
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    }
                 }
             }
         },